Add NaN handling to loss calculations in SelfPredictiveDistanceModel and SelfPredictiveDistanceHLGModel
This commit introduces the use of jnp.nan_to_num to handle NaN, positive infinity, and negative infinity values in the loss calculations of the SelfPredictiveDistanceModel and SelfPredictiveDistanceHLGModel classes. This enhancement aims to improve numerical stability during training by ensuring that loss values remain within a defined range.

+++ neural_util/modules.py
@@ -6,7 +6,7 @@ import jax.numpy as jnp
 
 from .norm import BatchReNorm, DyTan
 
-DTYPE = jnp.float32
+DTYPE = jnp.bfloat16
 # Use float32 for numerically sensitive heads / losses.
 HEAD_DTYPE = jnp.float32
